Remove interfaces
